pyproject.toml
[project]
name = "ark-agentic"
version = "0.5.8"
description = "Lightweight ReAct agent framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
    "flashtext>=2.7",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "openinference-semantic-conventions",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "bcrypt>=5.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "apscheduler>=3.11",
    "opentelemetry-exporter-otlp-proto-http",
    "openinference-instrumentation-langchain",
    "arize-phoenix-otel",
]
# db: 切换到 PostgreSQL 时加装（默认 SQLite 不需要）
db = [
    "psycopg[binary]>=3.2.0",
]
pa-jt = [
    "pycryptodome>=3.20.0",
]

[project.scripts]
ark-agentic = "ark_agentic.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ark_agentic"]
exclude = [
    "src/ark_agentic/agents/insurance",
    "src/ark_agentic/agents/securities",
    "src/ark_agentic/app.py",
    "src/ark_agentic/portal",
    "src/ark_agentic/plugins/studio/frontend/node_modules",
    "src/ark_agentic/plugins/studio/frontend/src",
]

[tool.hatch.build.targets.sdist.force-include]
"src/ark_agentic/plugins/studio/frontend/dist" = "src/ark_agentic/plugins/studio/frontend/dist"

[tool.hatch.build.targets.wheel.force-include]
"src/ark_agentic/plugins/studio/frontend/dist" = "ark_agentic/plugins/studio/frontend/dist"
"src/ark_agentic/agents/meta_builder/skills"   = "ark_agentic/agents/meta_builder/skills"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
# 防止单个用例死循环/阻塞 IO 拖死「全量测试」；真慢用例应拆 integration 或 mock
timeout = 180
markers = [
    "slow: 故意慢的集成测试；快速门禁可加 -m \"not slow\"",
]

[tool.pyright]
extraPaths = ["src"]

[tool.uv]
default-groups = ["dev"]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "twine>=6.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.3.0",
    "jsonschema>=4.26.0",
    "ark-agentic[server,db,pa-jt]",
]
securities = [
    "akshare>=1.18.43",
    "rapidfuzz>=3.14.3",
    "pypinyin>=0.55.0",
]
